refactor(avatar): centraliser le style des avatars dans useAvatar
Extrait la logique dupliquée (couleur, initiales) de 9 composants
vers un composable unique pour uniformiser le rendu des avatars.

🪄 Commit via GitWand
